requirements.txt
requests==2.27.1
beautifulsoup4==4.11.1
python-telegram-bot==20.0a0
